requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.2
annotated-types==0.7.0
attrs==25.3.0
av==14.4.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comfyui-embedded-docs==0.2.3
comfyui-frontend-package==1.23.4
comfyui-workflow-templates==0.1.30
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
cython==3.1.2
deprecated==1.2.18
diffusers==0.34.0
easydict==1.13
einops==0.8.1
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.2.3
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
insightface==0.7.3
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kornia==0.8.1
kornia-rs==0.1.9
lazy-loader==0.4
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matrix-client==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.1
mypy-extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
nunchaku @ https://github.com/mit-han-lab/nunchaku/releases/download/v0.3.1/nunchaku-0.3.1+torch2.7-cp312-cp312-win_amd64.whl
onnx==1.18.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygit2==1.18.0
pygithub==2.6.1
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.5.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
spandrel==0.4.1
sqlalchemy==2.0.41
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.16
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torchsde==0.2.6
tqdm==4.67.1
trampoline==0.1.2
transformers==4.52.4
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==1.26.20
uv==0.7.17
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
customtkinter
